Fixes #12179 - extracting strings in  host_inherited_params correctly
